+++ Cargo.toml
@@ -7,8 +7,6 @@ edition = "2021"
 
 [dependencies]
 bevy = "0.12.0"
-bevy-inspector-egui = "0.21.0"
-bevy_egui = "0.23.0"
 bevy_rand = "0.4.0"
 bevy_third_person_camera = "0.1.6"
 bevy_xpbd_3d = { version = "0.3.2", default-features = false, features = ["3d", "f32", "debug-plugin", "collider-from-mesh", "async-collider", "simd", "parallel"] }
