+++ src/alien/alien_plugin.rs
@@ -1,6 +1,7 @@
 use bevy::app::{App, Plugin, Update};
 use bevy::prelude::{in_state, IntoSystemConfigs, OnEnter};
-use crate::alien::systems::spawn_aliens::{alien_spawner_system, start_some_animations, load_animations, spawn_aliens};
+use crate::alien::systems::spawn_aliens::{alien_spawner_system, spawn_aliens};
+use crate::animation::animation_plugin::{load_animations, start_some_animations};
 use crate::game_state::GameState;
 
 pub struct AlienPlugin;
@@ -24,15 +25,10 @@ impl Plugin for StatefulAlienPlugin {
     fn build(&self, app: &mut App) {
         app
             .add_systems(
-                OnEnter(GameState::InGame),
-                load_animations,
-            )
-            .add_systems(
                 Update,
                 (
                     alien_spawner_system,
                     spawn_aliens,
-                    start_some_animations
                 ).run_if(in_state(GameState::InGame)),
             );
     }
